Some more warnings suppressed
git-svn-id: https://svn.aros.org/svn/aros/trunk@12643 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ workbench/tools/commodities/AutoPoint.c
@@ -191,7 +191,7 @@ static BOOL initiate(int argc, char **argv, APState *as)
 	if (*(struct Library **)tmpLibTable->lT_Library == NULL)
 	{
 	    printf("%s %s %i\n", getCatalog(catalogPtr, MSG_CANT_OPEN_LIB),
-		   tmpLibTable->lT_Name, tmpLibTable->lT_Version);
+		   tmpLibTable->lT_Name, (int)tmpLibTable->lT_Version);
 
 	    return FALSE;
 	}
